requirements.txt
Django==1.4.2
RapidSMS==0.13.0
South==0.7.6
Unidecode==0.04.12
amqp==1.0.11
anyjson==0.3.3
billiard==2.7.3.28
celery==3.0.19
celery-with-redis==3.0
distribute==0.6.28
django-celery==3.0.17
django-common-helpers==0.3
-e git://github.com/mvpdev/django-eav.git@24be357a3044a7d5aa0a3ae482766c158fd59dce#egg=django_eav-dev
django-nose==1.1
django-picklefield==0.2.1
django-redis-cache==0.9.7
django-selectable==0.7.0
django-tables2==0.13.0
django-tastypie==0.9.14
djappsettings==0.1.0
djtables==0.1.2
flower==0.5.0
httplib2==0.8
kombu==2.5.10
mimeparse==0.1.3
nose==1.2.1
pysqlite==2.6.3
python-dateutil==1.5
pytz==2013b
raven==3.3.7
readline==6.2.4.1
redis==2.7.3
requests==1.2.0
sphinx-me==0.1.2
tornado==3.0.1
wsgiref==0.1.2
